Removed tests targeting the request writer specifically as they provide no additional coverage, refactored start response parsing into its own static class
