--- NSonic.Tests/SonicRequestWriterAsyncTests.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace NSonic.Tests
-{
-    [TestClass]
-    public class SonicRequestWriterAsyncTests : SonicRequestWriterTests
-    {
-        protected override bool Async => true;
-    }
-}
